Common: move EnrichmentConf into its own module (close #303)
